+++ RELEASENOTES.md
@@ -23,6 +23,9 @@
         `MediaSource.Factory.setLoadErrorHandlingPolicy`. Instances of
         `DefaultDrmSessionManagerProvider` and `DefaultLoadErrorHandlingPolicy`
         can be passed explicitly if required.
+    *   Add `MediaItem.RequestMetadata` to represent metadata needed to play
+        media when the exact `LocalConfiguration` is not known. Also remove
+        `MediaMetadata.mediaUrl` as this is now included in `RequestMetadata`.
 *   Track selection:
     *   Flatten `TrackSelectionOverrides` class into `TrackSelectionParameters`,
         and promote `TrackSelectionOverride` to a top level class.